[update] webhook

+++ SmartFarmManager.API/Controllers/FarmController.cs
@@ -48,7 +48,7 @@ namespace SmartFarmManager.API.Controllers
                 {
                     Name = request.Name,
                     Address = request.Address,
-                    Area = request.Area,
+                    Area = (double)request.Area,
                     PhoneNumber = request.PhoneNumber,
                     Email = request.Email
                 });
@@ -137,7 +137,7 @@ namespace SmartFarmManager.API.Controllers
                 {
                     Name = request.Name,
                     Address = request.Address,
-                    Area = request.Area,
+                    Area =(double) request.Area,
                     PhoneNumber = request.PhoneNumber,
                     Email = request.Email
                 });
